requirements.txt
pandas
numpy
fastapi
uvicorn
sqlalchemy
google-generativeai
python-dotenv
pydantic
